requirements.txt
Django==1.6.5
South==1.0
django-extensions==1.3.10
djangorestframework==2.3.14
drf-extensions==0.2.5

# debug:
Werkzeug==0.9.6
django-debug-toolbar==1.2.1
